Misc fixes from rebasing
